refactored logic/cmp operations location
