+++ core/src/interpreter/mod.rs
@@ -192,85 +192,6 @@ impl From<InterpretedType> for bool {
     }
 }
 
-impl TermNode {
-    fn arith_op(
-        &self,
-        arith_type: char,
-        lhs: InterpretedType,
-        rhs: InterpretedType,
-        context: &ExecutionContext,
-    ) -> InterpreterResult {
-        match arith_type {
-            '+' => Ok(lhs + rhs),
-            '-' => Ok(lhs - rhs),
-            '*' => Ok(lhs * rhs),
-            '/' => match lhs / rhs {
-                Ok(res) => Ok(res),
-                Err(err) => Err(err.into(context)),
-            },
-            _ => unreachable!("unknown arithmetic operation"),
-        }
-    }
-
-    fn comp_op(
-        &self,
-        cmp_type: CompType,
-        lhs: InterpretedType,
-        rhs: InterpretedType,
-    ) -> InterpreterResult {
-        match cmp_type {
-            CompType::EE => Ok(InterpretedType::Int(i64::from(lhs == rhs))),
-            CompType::NE => Ok(InterpretedType::Int(i64::from(lhs != rhs))),
-            CompType::GT => Ok(InterpretedType::Int(i64::from(lhs > rhs))),
-            CompType::GTE => Ok(InterpretedType::Int(i64::from(lhs >= rhs))),
-            CompType::LT => Ok(InterpretedType::Int(i64::from(lhs < rhs))),
-            CompType::LTE => Ok(InterpretedType::Int(i64::from(lhs <= rhs))),
-        }
-    }
-
-    fn logic_op(
-        &self,
-        lgc_type: LogicType,
-        lhs: bool,
-        rhs: bool,
-    ) -> InterpreterResult {
-        match lgc_type {
-            LogicType::AND => {
-                Ok(InterpretedType::Int(i64::from(lhs && rhs)))
-            },
-            LogicType::OR => {
-                Ok(InterpretedType::Int(i64::from(lhs || rhs)))
-            },
-            _ => unreachable!("Cannot interpret ! operation for a term")
-        }
-    }
-}
-
-impl UnaryNode {
-    fn arith_op(&self, arith_type: char, context: &mut ExecutionContext) -> InterpreterResult {
-        match arith_type {
-            '+' => self.node.interpret(context),
-            '-' => {
-                let rhs = self.node.interpret(context)?;
-                context.current_pos = self.pos;
-                Ok(InterpretedType::Int(-1) * rhs)
-            }
-            _ => unreachable!("A unary arithmetic operation can only be -/+"),
-        }
-    }
-
-    fn logic_op(&self, lgc_type: LogicType, context: &mut ExecutionContext) -> InterpreterResult {
-        match lgc_type {
-            LogicType::NOT => {
-                let rhs = self.node.interpret(context)?;
-                context.current_pos = self.pos;
-                Ok(InterpretedType::Int(i64::from(!bool::from(rhs))))
-            }
-            _ => unreachable!("A unary operator can only be !"),
-        }
-    }
-}
-
 pub trait Interpret<'a> {
     // I should try to add some trait methods that "visits" the children nodes
     // and use those instead.
